Move methods to a new scrapy.utils.versions
