+++ scrapy/commands/version.py
@@ -2,7 +2,7 @@ from __future__ import print_function
 
 import scrapy
 from scrapy.commands import ScrapyCommand
-from scrapy.utils.log import scrapy_components_versions
+from scrapy.utils.versions import scrapy_components_versions
 
 
 class Command(ScrapyCommand):
